requirements.txt
numpy
pandas
sklearn
matplotlib
python-mnist
